% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{combine_hubs}
\alias{combine_hubs}
\title{Hub detection on a graph : three methods}
\usage{
combine_hubs(metric, R)
}
\arguments{
\item{metric}{chr weighting scheme like "FA", "GFA"}

\item{R}{is a list of graph containing G1, G2, G3 (one graph for each visit).}
}
\description{
Nodes are defined as Hubs if they verify the inequality :
X > mean(X) + sd(X)
Where X can be node strength, betweenness centrality, or closeness centrality.
The difference with the other hub detection function is that it is less restrictive. We consider a node as a hub
if it fullfill the inequality for at least one method. Here the function has been created to work on group average data (three graphs in input,
corresponding to each average graph for a visit)
}
\examples{
combine_hubs("FA",R)
}
