% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_local_analysis.R
\name{compute_on_all_nodes}
\alias{compute_on_all_nodes}
\title{Evaluate group influence every node-level topology measure : Multiple Kruskal-Wallis}
\usage{
compute_on_all_nodes(whole_data, metric, eval, thresh_method, tvalue)
}
\arguments{
\item{whole_data}{the data frame containing all data for one weighting scheme}

\item{metric}{chr the weighting scheme such as "FA", "GFA"}

\item{eval}{chr the global network topology measure : like clust_coeff, global_eff}

\item{thresh_method}{chr thresholding procedure, either "threshold" or "density"}

\item{tvalue}{float threshold value, either a threshold value or a density value}
}
\description{
Perform m Kruskal Wallis on the three groups, where m is the number of nodes. Dependent variable is the vertex-level topology measure.
within factor is the group (V1,V2,V3). You can choose a thresholding method : threshold or density, to perform the analysis on thresholded matrix.
}
\examples{
eval_on_single_threshold("FA","global_eff",0.2)
}
