% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_rich_club_analysis.R
\name{compute_phi_patients}
\alias{compute_phi_patients}
\title{Compute all phi for all subject in one group (v_id)}
\usage{
compute_phi_patients(data, v_id, metric, tvalue, k, nrand)
}
\arguments{
\item{data}{a dataframe containing all data for analysis}

\item{v_id}{visit id : "V1","V2","V3"}

\item{metric}{the metric chosen for the weight in the network. "FBC","FA"..}

\item{tvalue}{density threshold value. Default .15}
}
\description{
Compute all phi for all subject in one group (v_id)
}
