% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_rich_club_analysis.R
\name{computeGr}
\alias{computeGr}
\title{Construct a random graph from a graph g
Respect the same nb of edges/nodes. And keep the same degree distribution
and keep edge weight distribution}
\usage{
computeGr(g)
}
\arguments{
\item{g}{a undirect weighted graph}
}
\description{
Construct a random graph from a graph g
Respect the same nb of edges/nodes. And keep the same degree distribution
and keep edge weight distribution
}
