% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_rich_club_analysis.R
\name{computePhi}
\alias{computePhi}
\title{Compute non normalized rich club coefficient of a graph g}
\usage{
computePhi(g, k)
}
\arguments{
\item{g}{a undirect weighted graph}

\item{k}{the degree threshold in the induced subgraph construction}
}
\description{
Compute non normalized rich club coefficient of a graph g
}
