% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_rich_club_analysis.R
\name{computePhiGroup}
\alias{computePhiGroup}
\title{Do all statistical analysis at each k}
\usage{
computePhiGroup(dataFBC, metric, t, k)
}
\arguments{
\item{metric}{the metric chosen for the weight in the network. "FBC","FA"..}

\item{data}{a dataframe containing all data for analysis}

\item{tvalue}{density threshold value. Default .15}
}
\description{
Do all statistical analysis at each k
}
