% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_local_analysis.R
\name{computeSparseMetric}
\alias{computeSparseMetric}
\title{Compute the measure of choice for each node for each patient in a group ("V1", "V2", or "V3)}
\usage{
computeSparseMetric(
  data,
  v_id,
  WM_metric = c("FBC", "ODI", "GFA", "Fintra", "FA"),
  eval = c("degree", "clust_coeff", "strength", "betweenness", "closeness", "eigen",
    "efficiency"),
  thresh_method,
  tvalue
)
}
\arguments{
\item{data}{dataframe containing all information for one weighting scheme}

\item{v_id}{chr visit id, like "V1", "V3"}

\item{eval}{chr the measure to access the network topology}

\item{tvalue}{float either a threshold value or a density value}

\item{metric}{chr the weighting scheme}

\item{threhs_method}{chr thresh_method is either "threshold" or "density".}
}
\value{
vector of size n x m
}
\description{
From the dataframe containing all information for one weighting scheme, this function returns a vector
of n values by m nodes, where n is the number of participant in a group. And the values are global measure of network topology.
}
\examples{
computeSW(dataFA,"V1","FA","global_eff",0.2)
}
