% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_rmanova.R
\name{computeSW}
\alias{computeSW}
\title{Compute the measure of choice for each patient in a group ("V1", "V2", or "V3)}
\usage{
computeSW(
  data,
  v_id,
  WM_metric = c("FBC", "ODI", "GFA", "Fintra", "FA"),
  eval = c("clust_coeff", "characteristic_path", "global_eff", "local_eff",
    "smallworldeness", "richcore", "strength", "betweenness"),
  thresh_method,
  tvalue
)
}
\arguments{
\item{data}{dataframe containing all information for one weighting scheme}

\item{v_id}{chr visit id, like "V1", "V3"}

\item{eval}{chr the measure to access the network topology}

\item{metric}{chr the weighting scheme}

\item{threshold}{float a threshold if the data is computed on a specific threshold. default = 0}
}
\description{
From the dataframe containing all information for one weighting scheme, this function returns a vector
of n values, where n is the number of participant in a group. And the values are global measure of network topology.
}
\examples{
computeSW(dataFA,"V1","FA","global_eff",0.2)
}
