% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_local_analysis.R
\name{get_one_local_result}
\alias{get_one_local_result}
\title{This function compute metrics for one node in particular
It need the main_group_local_analysis function to have run}
\usage{
get_one_local_result(df, L1, L2, L3, node)
}
\arguments{
\item{df}{dataframe containing all information for one weighting scheme}

\item{L1}{matrix containing as column as subjects, as rows as nodes. L1 is for V1}

\item{node}{str containing the name of the node in the atlas}
}
\description{
This function compute metrics for one node in particular
It need the main_group_local_analysis function to have run
}
