% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plot_local_results.R
\name{get_subcortical_labels}
\alias{get_subcortical_labels}
\title{Table of region names equivalence in the subcortical region}
\usage{
get_subcortical_labels()
}
\value{
list of two tables for left and right hemisphere. tables of equivalence
}
\description{
Create a equivalence table between region names of destrieux atlas in the package ggseg, and region names used in freesurfer.For subcortical regions
}
