% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_load_data.R
\name{get_subject_ids}
\alias{get_subject_ids}
\title{Get the list of subject participating visit i}
\usage{
get_subject_ids(data, v_id)
}
\arguments{
\item{data}{dataframe (created by 'read_and_normalize')}

\item{v_id}{chr visit id, like "V1", "V2", "V3".}
}
\value{
list
}
\description{
Get the list of subject participating visit i
}
