% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_linear_model.R
\name{getData}
\alias{getData}
\title{Group all measures by group, and normalize to normal distribution}
\usage{
getData(data, WM_metric, eval, thresh_method, tvalue)
}
\arguments{
\item{data}{dataframe containing all information for one weighting scheme}

\item{eval}{chr the measure to access the network topology}

\item{metric}{chr the weighting scheme}

\item{threshold}{float a threshold if the data is computed on a specific threshold. default = 0}
}
\description{
Compute each measure of network topology for each visit group (V1,V2,V3). Group this into
a two columns data frame : group, y. Finally, use bestNormalize to normalize y, to ensure that
the normality hypothesis is verified in Anova
}
