% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_load_data.R
\name{getLUT}
\alias{getLUT}
\title{Create a dataframe from freesurfer lut.}
\usage{
getLUT()
}
\value{
dataframe of two columns : No for the region id, labelname for the region label
}
\description{
Create a dataframe from freesurfer lut.
}
\examples{
getLUT()
}
