% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{getMeanSparsity}
\alias{getMeanSparsity}
\title{Return the minimum graph density across all participants and all visits}
\usage{
getMeanSparsity(data, WM_metric = c("FBC", "ODI", "GFA", "FA", "Fintra"))
}
\arguments{
\item{data}{dataframe containing all data for one weighting scheme}

\item{metric}{chr the weighting scheme : "FA", "FBC", "GFA"...}
}
\value{
float group minimum density
}
\description{
Compute all graph densities and returns the mininum
}
\examples{
getMeanSparsity(dataFA,"FA")
}
