% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_rmanova.R
\name{getNormalizeData}
\alias{getNormalizeData}
\title{Add normalized data column to the dataframe of the results}
\usage{
getNormalizeData(Gdata, listScores)
}
\arguments{
\item{Gdata}{dataframe of patients ids, timepoint (group) and score}

\item{listScores}{list of normalized scores with respect to V1}
}
\value{
Gdata with an additional column of normalized data
}
\description{
Add normalized data column to the dataframe of the results
}
