% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_linear_model.R
\name{getResultsInNetworks}
\alias{getResultsInNetworks}
\title{Group all measures by group. Generalization of getData (to test)}
\usage{
getResultsInNetworks(
  DF,
  WM_metric,
  eval,
  thresh_method,
  threshold,
  rsnet,
  groups
)
}
\arguments{
\item{eval}{chr the measure to access the network topology}

\item{threshold}{float a threshold if the data is computed on a specific threshold. default = 0}

\item{data}{dataframe containing all information for one weighting scheme}

\item{metric}{chr the weighting scheme}
}
\description{
Compute each measure of network topology for each visit group (V1,V2,V3). Group this into
a two columns data frame : group, y. Finally, use bestNormalize to normalize y, to ensure that
the normality hypothesis is verified in Anova
}
