% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{hub_detectionh}
\alias{hub_detectionh}
\title{Hub detection on a graph using different methods}
\usage{
hub_detectionh(g, method, data_path)
}
\arguments{
\item{g}{igraph object}

\item{method}{chr method for hub detection "closeness","strength","betweenness"}
}
\description{
Nodes are defined as Hubs if they verify the inequality :
X > mean(X) + sd(X)
Where X can be node strength, betweenness centrality, or closeness centrality
Provide a horizontal histogram plot
}
