% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_visualize_on_interface.R
\name{interfaceConhect}
\alias{interfaceConhect}
\title{Generate an interface to plot graph for a patient & visit & weighting scheme.}
\usage{
interfaceConhect(metric)
}
\arguments{
\item{metric}{a chr for the weighting scheme : "FA", "FBC", "GFA"...}
}
\value{
a shiny interface
}
\description{
Generate an interface to plot graph for a patient & visit & weighting scheme.
}
\examples{
interfaceConhect("FA")
}
