% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_local_analysis.R
\name{isnormal}
\alias{isnormal}
\title{Shapiro Wilk test for normality}
\usage{
isnormal(Gdata)
}
\arguments{
\item{Gdata}{dataframe two columns : group, y, created by getData}
}
\description{
Perform a Shapiro Wilk test to test the normality hypothesis. To further
perform parametric test Anova.
}
\examples{
isnormal(GdataFA)
}
