% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plot_local_results.R
\name{join_res_sub_labels}
\alias{join_res_sub_labels}
\title{Joint of results and region names. subcortical regions}
\usage{
join_res_sub_labels(res_cort, atlas_subcort)
}
\arguments{
\item{atlas_subcort}{dataframe of the tables of equivalence for subcortical regions}

\item{rest_cort}{dataframe created in split_zone_hemi}
}
\value{
list of two tables for left and right hemisphere. tables of equivalence
}
\description{
Join the two tables by the freesurfer name (unique key), because ggseg cannot plot with freesurfer names
}
