% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_linear_model.R
\name{LinearModel_analysis}
\alias{LinearModel_analysis}
\title{Linear Mixed Model}
\usage{
LinearModel_analysis(Gdata)
}
\arguments{
\item{Gdata}{dataframe two columns : group, y, created by getData}
}
\description{
Perform a RMANOVA on the three groups. Dependent variable is the topology measure.
within factor is the group (V1,V2,V3).
}
