% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plot_local_results.R
\name{main_cortical_p}
\alias{main_cortical_p}
\title{Plot Kruskal Wallis result on cortical regions}
\usage{
main_cortical_p(metric, eval, thresh_method, tvalue)
}
\arguments{
\item{metric}{chr the weighting scheme such as "FA", "GFA"}

\item{eval}{chr the global network topology measure : like clust_coeff, global_eff}

\item{thresh_method}{chr the method of threshoding. Whether "threshol" or "density"}

\item{threshold}{float the threshold/density to be applied on each network in the analysis.}
}
\value{
two figures : cortical and subcortical figures
}
\description{
Use split_zone_hemi, get_(sub)cortical_labels, to get the values to plot. Use ggseg to plot the values on the destrieux atlas.
Plot only V1 -V2 and V1 - V3
}
