% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_linear_model.R
\name{main_global_llm}
\alias{main_global_llm}
\title{Evaluate group influence on one global network topology measure : Linear model}
\usage{
main_global_llm(
  data_path,
  weighting_scheme = c("GFA", "FBC", "FA", "Fintra", "ODI"),
  eval = c("clust_coeff", "characteristic_path", "global_eff", "local_eff",
    "smallworldeness", "richcore", "strength", "betweenness"),
  thresh_method,
  tvalue
)
}
\arguments{
\item{weighting_scheme}{chr the weighting scheme such as "FA", "GFA"}

\item{eval}{chr the global network topology measure : like clust_coeff, global_eff}

\item{thresh_method}{chr the method of threshoding. Whether "threshold" or "density"}

\item{threshold}{float the threshold/density to be applied on each network in the analysis.}
}
\description{
Perform a RMANOVA on the three groups. Dependent variable is the topology measure.
within factor is the group (V1,V2,V3). You can choose a threshold to perform the analysis on thresholded matrix. default is 0
}
