% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_local_analysis.R
\name{main_group_local_metrics_analysis}
\alias{main_group_local_metrics_analysis}
\title{Compute the local measure of choice for each patient in a group ("V1", "V2", or "V3) and perform the group level statistical analysis}
\usage{
main_group_local_metrics_analysis(
  df,
  group_list,
  WM_metric,
  eval = c("degree", "clust_coeff", "strength", "betweenness", "closeness", "eigen",
    "efficiency"),
  thresh_method,
  tvalue
)
}
\arguments{
\item{df}{dataframe containing all information for one weighting scheme}

\item{WM_metric}{chr the weighting scheme}

\item{eval}{chr the measure to access the network topology}

\item{thresh_method}{either "threshold" for absolute thresholding, or "density" for density thresholding}

\item{tvalue}{either the value of the absolute threshold or density threshold}

\item{group_mist}{chr visit id, like "V1", "V3"}
}
\description{
Fit a linear mixed model for each region of interest (local metric as dependent variable, group as independant variable, and subject id as random effect to account for repetitive data)
Use correction for multiple comparison (bonferronni, fdr)
}
