% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_local_analysis.R
\name{main_local_analysis}
\alias{main_local_analysis}
\title{Evaluate group influence every node-level topology measure : Multiple Kruskal-Wallis}
\usage{
main_local_analysis(metric, eval, thresh_method, tvalue)
}
\arguments{
\item{metric}{chr the weighting scheme such as "FA", "GFA"}

\item{eval}{chr the global network topology measure : like clust_coeff, global_eff}

\item{thresh_method}{chr thresholding procedure, either "threshold" or "density"}

\item{tvalue}{float threshold value, either a threshold value or a density value}

\item{whole_data}{the data frame containing all data for one weighting scheme}
}
\value{
results, a plot on a heatmap, a plot on the brain atlases
}
\description{
Perform m Kruskal Wallis on the three groups, where m is the number of nodes. Dependent variable is the vertex-level topology measure.
within factor is the group (V1,V2,V3). You can choose a thresholding method : threshold or density, to perform the analysis on thresholded matrix.
}
\examples{
main_local_analysis("FA","global_eff",0.2)
}
