% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_rich_club_analysis.R
\name{main_rich_club_analysis}
\alias{main_rich_club_analysis}
\title{Main function for the rich club analysis}
\usage{
main_rich_club_analysis(
  data,
  metric = "FBC",
  tvalue = 0.15,
  kmin = 10,
  kmax = 50
)
}
\arguments{
\item{data}{a dataframe containing all data for analysis}

\item{metric}{the metric chosen for the weight in the network. "FBC","FA"..}

\item{tvalue}{density threshold value. Default .15}

\item{kmin}{minimum degree for the range of analysis default 10}

\item{kmax}{maximum degree for the range of analysis default 50}
}
\description{
Main function for the rich club analysis
}
