% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{makeGraph}
\alias{makeGraph}
\title{Create a igraph object representing a subject at one visit, form one weighting scheme, at a given threshold.}
\usage{
makeGraph(
  data,
  s_id,
  v_id,
  WM_metric = c("FBC", "ODI", "GFA", "FA", "Fintra", "PearsonCorrel"),
  threshold = 0
)
}
\arguments{
\item{data}{dataframe containing all data for one weighting scheme}

\item{s_id}{chr the subject id: like "1-101-MG"...}

\item{v_id}{chr the visit id like "V1"...}

\item{threshold}{float a threshold value, default = 0}

\item{metric}{chr the weighting scheme : "FA", "FBC", "GFA"...}
}
\value{
a igraph object
}
\description{
Create a weighted undirected graph with igraph. The graph represent a patient (ex : "1-E-101-MG"), at one visit (ex : "V1"),
for one weighting scheme on the edges (ex : "FA"). You can as well choose a threshold value. This will truncate the graph by
deleting the edges that have a value below the threshold. Default is 0
}
