% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{mindensity}
\alias{mindensity}
\title{Compute the maximal density to use as a threshold}
\usage{
mindensity(df, groups, WM_metric, rsnet = "All")
}
\arguments{
\item{groups}{list of groups (ex : "V1","V2")}

\item{WM_metric}{str metric of white matter for edge weighting}

\item{g}{igraph object}
}
\description{
It is computed as the minimal density of graphs in the dataset
}
