% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{module_detection}
\alias{module_detection}
\title{Identify modules of a graph and membership of each node}
\usage{
module_detection(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
Module detection is done with Louvain algorithm, for cluster detection.
}
