% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{normalized_shortest_path}
\alias{normalized_shortest_path}
\title{Compute the normalized characteristic path lengthof a graph}
\usage{
normalized_shortest_path(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
Compute the normalized characteristic path length of a graph.Normalized characteristic path length is defined by dividing characteristic path length
by lambda' which is the mean of values obtain from 100 random graphs (with same number of edges/nodes and degree/weight distribution)
}
