% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{phinorm}
\alias{phinorm}
\title{Compute the phinorm coefficient for rich club analysis}
\usage{
phinorm(g, k)
}
\arguments{
\item{g}{igraph object}

\item{k}{int}
}
\description{
phi norm is defined as the ratio of phi(k) (the density of a graph with nodes of degree >k), divided by the phi_rand(k),
which is the same measure for a random graph (with same number of edges/nodes and degree/weight distribution)
}
\examples{
phinorm(g,10)
}
