% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{plot_adjacency}
\alias{plot_adjacency}
\title{Plot the adjacency matric of a graph}
\usage{
plot_adjacency(g, WM_metric = c("FBC", "ODI", "GFA", "FA", "Fintra"))
}
\arguments{
\item{g}{igraph object}

\item{metric}{chr the weighting scheme : "FA", "FBC", "GFA"...}
}
\description{
This function plot the adjacency matrix of a graph as a heatmap.
}
