% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_visualisation_tools_brain.R
\name{plot_cortdesterieux2d}
\alias{plot_cortdesterieux2d}
\title{Plot Desterieux (freesurfer) cortical atals in 2d with labels}
\usage{
plot_cortdesterieux2d()
}
\description{
Plot Desterieux (freesurfer) cortical atals in 2d with labels
}
