% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_visualisation_tools_brain.R
\name{plot_dataonbrain}
\alias{plot_dataonbrain}
\title{Plot data on brain on an indiviudal level. On destrieux atlas. from data giving each cortical region value.}
\usage{
plot_dataonbrain(dataOnBrain)
}
\arguments{
\item{dataOnBrain}{}
}
\description{
Plot data on brain on an indiviudal level. On destrieux atlas. from data giving each cortical region value.
}
