% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_linear_model.R
\name{plot_dataviz}
\alias{plot_dataviz}
\title{plot the data for a good visualization}
\usage{
plot_dataviz(G, lab)
}
\arguments{
\item{G}{the dataframe}
}
\description{
plot the data for a good visualization
}
