% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plot_local_results.R
\name{plot_p_on_brain_map}
\alias{plot_p_on_brain_map}
\title{plot significant values on brain map}
\usage{
plot_p_on_brain_map(metric, eval, thresh_method, tvalue)
}
\description{
plot significant values on brain map
}
