% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{plot_ridge_distribution}
\alias{plot_ridge_distribution}
\title{Plot the edges weight distribution for each participant of a group. FA weighted}
\usage{
plot_ridge_distribution(dataFA, v_id)
}
\arguments{
\item{dataFA}{dataframe containing all the FA weighted data}

\item{v_id}{chr visit id like "V1", "V2"...}
}
\description{
Plot the fractional anistropy edges weight distribution each participants of a group
(V1, V2, V3). Plot it as a ridge plot. Note : You can plot the three ridges side to side
by calling this function three times and using gridExtra::grid.arrange.
}
