% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{plot_weight_distribution}
\alias{plot_weight_distribution}
\title{Plot the edges weight distribution for one participant : FBC}
\usage{
plot_weight_distribution(data)
}
\arguments{
\item{data}{edge list}
}
\description{
TO DELETE OR UPDATE
Plot the fiber count edges weight distribution on a log scale for one participant.
}
\examples{
plot_weight_distribution(data)
}
