% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{plot_weight_distribution_graph}
\alias{plot_weight_distribution_graph}
\title{Plot the edges weight distribution for one participant : FBC}
\usage{
plot_weight_distribution_graph(g, nBreaks)
}
\arguments{
\item{g}{igraph object}

\item{nBreaks}{number of breaks for the histogram}
}
\description{
TO DELETE OR UPDATE
Plot the fiber count edges weight distribution on a log scale for one participant. from a igraph object
}
\examples{
plot_weight_distribution_graph(g,40)
}
