% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{plot_weight_distribution_graph_fa}
\alias{plot_weight_distribution_graph_fa}
\title{Plot the edges weight distribution for one participant : FA}
\usage{
plot_weight_distribution_graph_fa(g, v_id, nBreaks)
}
\arguments{
\item{g}{igraph object}

\item{nBreaks}{number of breaks for the histogram}
}
\description{
TO DELETE OR UPDATE
Plot the fractional anistropy edges weight distribution for one participant.from a graph
}
\examples{
plot_weight_distribution(data)
}
