% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_local_analysis.R
\name{plotSaveLocalHeatmap}
\alias{plotSaveLocalHeatmap}
\title{Plot significant result as a heatmap}
\usage{
plotSaveLocalHeatmap(data, metric, eval, thresh_method, tvalue)
}
\arguments{
\item{data}{dataframe containing pvalues and either -1,0,1 to indicate if there is an increase, nothing or a decrease;}

\item{metric}{chr weighting scheme : "FA", "FBC"...}

\item{eval}{chr vertex level measure :"clust_coeff", "efficiency"...}

\item{thresh_method}{chr thresholding procedure, either "threshold" or "density"}

\item{tvalue}{float threshold value, either a threshold value or a density value}
}
\value{
save image in Results/metric/Local
}
\description{
Plot the post hoc result as a heatmap for V1-V2 and V1-V3.
light blue represent a significant decrease between V1 and V2/3.
regular blue represent non significant result
dark blue represent significant increase between V1 and V2/3
}
