% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{random_walk_length}
\alias{random_walk_length}
\title{Function to perform a random walk on a weighted graph with inverted weights}
\usage{
random_walk_length(g)
}
\arguments{
\item{g}{the weighted graph}
}
\description{
Function to perform a random walk on a weighted graph with inverted weights
}
