% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_load_data.R
\name{read_and_normalize_data}
\alias{read_and_normalize_data}
\title{#' Update a dataframe containing a column ROI1_ROI2 to two columns ROI1 and ROI2.
#'
#' @param data the dataframe
#' @returns a data frame with the two additional columns
#' @export
separateRois <- function(data){}
\usage{
read_and_normalize_data(data_path, WM_metric)
}
\arguments{
\item{data_path}{chr indicating the path of the "raw" data}

\item{metric}{a chr for the weighting scheme : "FA", "FBC", "GFA"...}
}
\value{
the newly created dataframe
}
\description{

}
