% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_global_rmanova.R
\name{RMANOVA_analysis}
\alias{RMANOVA_analysis}
\title{Repeated Measure Anova}
\usage{
RMANOVA_analysis(Gdata)
}
\arguments{
\item{Gdata}{dataframe two columns : group, y, created by getData}
}
\description{
Perform a RMANOVA on the three groups. Dependent variable is the topology measure.
within factor is the group (V1,V2,V3).
}
\examples{
RMANOVA_analysis(GdataFA)
}
