% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_compute_metrics.R
\name{smallworldeness}
\alias{smallworldeness}
\title{Compute the smallworldenss of a graph}
\usage{
smallworldeness(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
Compute the smallworldeness of a graph in the sense of Hagmann 2008.
Smallworldeness is defined as the ratio of the normalized clustering by the normalized characteristic path length;
Normalized X is defined by dividing X by X' which is the mean of values obtain from 100 random graphs (with same number of edges/nodes and degree/weight distribution)
}
\examples{
smallworldeness(g)
}
