% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graph_construction.R
\name{sparseThresh}
\alias{sparseThresh}
\title{Create a igraph object at a specific wanted density}
\usage{
sparseThresh(g, wantedDensity)
}
\arguments{
\item{g}{igraph object}

\item{wantedDensity}{the density wanted for the new graph}
}
\value{
a igraph object at density 'wantedDensity'
}
\description{
Graph density is defined as the number of edges divided by the possible number of edges (defined as 0.5*nn(nn-1) where nn is the number of nodes).
The function calculates the number of edges that is necessary to delete to obtain the wantedDensity, delete the lowest weighted number of edges, and returns
a newly created graph, that has the wanted density.
}
\examples{
sparseThresh(sample_gnm(10,40),0.2)
}
