% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plot_local_results.R
\name{split_zone_hemi}
\alias{split_zone_hemi}
\title{Split zones in the brain : left/rigth hemisphere, and cortical/subcortical}
\usage{
split_zone_hemi(csv_res)
}
\arguments{
\item{csv_res}{a dataframe with columns X, Y , Z, coming from local analysis}
}
\value{
list of sub dataframe, for each zone
}
\description{
Split zones in the brain : left/rigth hemisphere, and cortical/subcortical
}
