% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_local_analysis.R
\name{computeSparseMetric}
\alias{computeSparseMetric}
\title{Compute the local measure of choice for each patient in a group ("V1", "V2", or "V3)
returns a list of list.(list of each n node local metric for each subject)}
\usage{
computeSparseMetric(
  data,
  v_id,
  WM_metric,
  eval = c("degree", "clust_coeff", "strength", "betweenness", "closeness", "eigen",
    "efficiency"),
  thresh_method,
  tvalue
)
}
\arguments{
\item{data}{dataframe containing all information for one weighting scheme}

\item{v_id}{chr visit id, like "V1", "V3"}

\item{WM_metric}{chr the weighting scheme}

\item{eval}{chr the measure to access the network topology}

\item{thresh_method}{either "threshold" for absolut thresholding, or "density" for density thresholding}

\item{tvalue}{either the value of the absolute threshold or density threshold}
}
\description{
Compute the local measure of choice for each patient in a group ("V1", "V2", or "V3)
returns a list of list.(list of each n node local metric for each subject)
}
