% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_load_data.R
\name{getDataDir}
\alias{getDataDir}
\title{Find the path containing the .xlsx file of the asked weighting scheme.}
\usage{
getDataDir(choice = c("FBC", "ODI", "GFA", "FA", "Fintra"))
}
\arguments{
\item{choice}{a chr for the weighting scheme : "FA", "FBC", "GFA"...}
}
\value{
chr the path containing the .xlsx file
}
\description{
Find the path containing the .xlsx file of the asked weighting scheme.
}
\examples{
getDataDir("FA")
}
